creado GameScore.kt
